backend/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
aspose-words==24.4.0
azure-common==1.1.28
azure-core==1.30.1
azure-mgmt-core==1.4.0
azure-mgmt-storage==21.1.0
azure-storage-blob==12.19.1
azure-storage-file-datalake==12.14.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
fastapi==0.115.11
Flask==2.2.5
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
lxml==5.1.0
MarkupSafe==2.1.5
nltk==3.8.1
openai==1.13.3
pycparser==2.21
pydantic==1.10.13
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
readability==0.3.1
regex==2023.12.25
requests==2.31.0
requests-mock==1.11.0
six==1.16.0
sniffio==1.3.1
tenacity==8.2.3
typing_extensions==4.10.0
urllib3==1.25.11
Werkzeug==2.2.3
aiohttp==3.9.3
clyent==1.2.1
